Support static methods, add more examples, only run callables
More people understand @staticmethod, so it makes sense to support this
and have it be default. Had to move the test checks out of a class,
because python2 does not preserve the @staticmethod decorator.

Also fixes a bug where it tried to call all public attributes (no good
if you add a class attribute).